Fixed bug with multiple views.
This is a good fix, but it could be better. Each Scintilla control has a
byte-to-char translation layer maintained in the LineCollection. But
when a single Document is shared between two Scintilla controls we end
up with two translation layers when we only need one. This fix will keep
them in sync, but a better fix would be to move the translation layer to
the Document object so that we only need one translation layer for
multiple views.