Fixes issue #87 and issue #94.
The issue was that Scintilla stores some properties (e.g. Encoding) with
the document. Changing to a new document caused those properties to be
reset. This fix is to carry those properties over to the new document. A
related issue was that the UndoCollection property would get reset and
cause the SCN_MODIFIED notification to return null for the modified
text. This would cause the byte-to-char layer to get out of sync. Since
the UndoCollection property is so dangerous I removed it altogether.