Added support for application idle time styling.
This commit also includes a new binary of the native SciLexer.dll
v3.6.3. The binary included in the last commit had the
Scintilla_DirectFunction function export mangled. I believe the reason
is because the Scintilla.def file was removed from the native build
process. This issue has been reported to Neil and bug 1813 opened to
track the issue. Until then I've had to modify the native scintilla.mak
file to include the module definition.
https://sourceforge.net/p/scintilla/bugs/1813/