Exposed the WordChars property. Issue #143.
I believe the reason I had this commented out before is because when you
call WordChars to get the default it included a bunch of non-printable
characters. My investigation suggests that there may be a bug in
Scintilla which includes everything between ASCII 127 and 255 as word
characters also, even though most people would not consider them as such
and Scintilla itself doesn't document this. To prevent people from
calling WordChars and getting 'junk' data, I've also added some code to
explicitly set the "word chars" to the values specified in Scintilla's
documentation.