Added ScintillaReader
Scintilla reader allows subsections of a Scintilla control to be
streamed using a buffered TextReader. This Makes it easier to use a
pre-existing lexer or parser since many of them are implemented in terms
of TextReader.

I'm assuming that its more efficient to get a few text ranges than many
individual characters and my default  buffer size is completely
arbitrary (256).